@@ -1,5 +1,5 @@
 {
-  "name": "my-app",
+  "name": "my-v0-project",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,30 +9,70 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "chart.js": "^4.4.0",
-    "cookie": "^1.0.2",
-    "d3-format": "^3.1.0",
-    "d3-time-format": "^4.1.0",
-    "leaflet": "1.9",
-    "next": "^15.2.4",
-    "oidc-client-ts": "^3.2.0",
-    "react": "^18.2.0",
-    "react-chartjs-2": "^5.3.0",
-    "react-datepicker": "^8.2.1",
-    "react-dom": "^18.2.0",
-    "react-financial-charts": "^2.0.1",
-    "react-leaflet": "^5.0.0-rc.1",
-    "recharts": "^2.1.10"
+    "@hookform/resolvers": "^3.9.1",
+    "@radix-ui/react-accordion": "1.2.2",
+    "@radix-ui/react-alert-dialog": "1.1.4",
+    "@radix-ui/react-aspect-ratio": "1.1.1",
+    "@radix-ui/react-avatar": "1.1.2",
+    "@radix-ui/react-checkbox": "1.1.3",
+    "@radix-ui/react-collapsible": "1.1.2",
+    "@radix-ui/react-context-menu": "2.2.4",
+    "@radix-ui/react-dialog": "1.1.4",
+    "@radix-ui/react-dropdown-menu": "2.1.4",
+    "@radix-ui/react-hover-card": "1.1.4",
+    "@radix-ui/react-label": "2.1.1",
+    "@radix-ui/react-menubar": "1.1.4",
+    "@radix-ui/react-navigation-menu": "1.2.3",
+    "@radix-ui/react-popover": "1.1.4",
+    "@radix-ui/react-progress": "1.1.1",
+    "@radix-ui/react-radio-group": "1.2.2",
+    "@radix-ui/react-scroll-area": "1.2.2",
+    "@radix-ui/react-select": "2.1.4",
+    "@radix-ui/react-separator": "1.1.1",
+    "@radix-ui/react-slider": "1.2.2",
+    "@radix-ui/react-slot": "1.1.1",
+    "@radix-ui/react-switch": "1.1.2",
+    "@radix-ui/react-tabs": "1.1.2",
+    "@radix-ui/react-toast": "1.2.4",
+    "@radix-ui/react-toggle": "1.1.1",
+    "@radix-ui/react-toggle-group": "1.1.1",
+    "@radix-ui/react-tooltip": "1.1.6",
+    "autoprefixer": "^10.4.20",
+    "chart.js": "latest",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "1.0.4",
+    "d3-time-format": "latest",
+    "date-fns": "4.1.0",
+    "embla-carousel-react": "8.5.1",
+    "input-otp": "1.4.1",
+    "leaflet": "latest",
+    "lucide-react": "^0.454.0",
+    "next": "15.2.4",
+    "next-themes": "^0.4.4",
+    "oidc-client-ts": "latest",
+    "react": "^19",
+    "react-chartjs-2": "latest",
+    "react-datepicker": "latest",
+    "react-day-picker": "8.10.1",
+    "react-dom": "^19",
+    "react-hook-form": "^7.54.1",
+    "react-leaflet": "latest",
+    "react-oidc-context": "latest",
+    "react-resizable-panels": "^2.1.7",
+    "recharts": "latest",
+    "sonner": "^1.7.1",
+    "tailwind-merge": "^2.5.5",
+    "tailwindcss-animate": "^1.0.7",
+    "vaul": "^0.9.6",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4",
-    "@types/chart.js": "^2.9.41",
-    "@types/d3-time-format": "^4.0.3",
-    "@types/leaflet": "^1.9.17",
-    "@types/node": "^20",
-    "@types/react": "^19.0.12",
-    "@types/react-dom": "^19.0.4",
-    "tailwindcss": "^4",
+    "@types/node": "^22",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "postcss": "^8.5",
+    "tailwindcss": "^3.4.17",
     "typescript": "^5"
   }
 }
